Addressing #185
Přesun testů do adresáře tests v rootu.

+++ tests/infofile_tests.py
@@ -9,30 +9,31 @@
 #-------------------------------------------------------------------------------
 
 import unittest
-import config, os
+import SharedTools.config, os
 
-infofile = config.RUIANDownloadInfoFile()
+infofile = SharedTools.config.RUIANDownloadInfoFile()
 
 class TestInfoFile(unittest.TestCase):
     FILENAME = "test.txt"
     LASTFULLDOWNLOAD_VALUE = "lastfulldownload"
     LASTPATCHDOWNLOAD_VALUE = "lastpatchdownload"
 
+
     def tearDown(self):
-        os.remove(self.FILENAME)
-        pass
+        if os.path.exists(self.FILENAME):
+            os.remove(self.FILENAME)
+
 
     def testSave(self):
         """ Tests Save and consequently Init """
-        f = infofile.InfoFile(self.FILENAME)
+        f = SharedTools.config.InfoFile(self.FILENAME)
         f.lastFullDownload = self.LASTFULLDOWNLOAD_VALUE
         f.lastPatchDownload = self.LASTPATCHDOWNLOAD_VALUE
         f.save()
 
-        f = infofile.InfoFile(self.FILENAME)
+        f = SharedTools.config.InfoFile(self.FILENAME)
         self.assertEqual(f.lastFullDownload,  self.LASTFULLDOWNLOAD_VALUE,  "lastFullDownload not read correctly")
         self.assertEqual(f.lastPatchDownload, self.LASTPATCHDOWNLOAD_VALUE, "lastPatchDownload not read correctly")
-        pass
 
 if __name__ == '__main__':
     unittest.main()
